pyproject.toml
[project]
name = "marker-server"
version = "0.1.0"
description = "Marker Server"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "pydantic==2.11.4",
    "pydantic-settings==2.9.1",
    "gunicorn==23.0.0",
    "fastapi==0.115.8",
    "python-multipart==0.0.20",
    "marker-pdf[full]==1.7.5",
    "uvicorn==0.34.0",
]

[project.optional-dependencies]
dev = [
    "ruff==0.6.5",
    "pre-commit==4.1.0",
]
test = [
    "pytest==8.3.4",
]

[tool.setuptools]
py-modules = []

[tool.ruff]
line-length = 150

[tool.ruff.lint]
ignore = ["F403", "F841"] # import * and never used variables

[tool.ruff.lint.isort]
force-sort-within-sections = true
known-first-party = ["config", "utils", "app"]
forced-separate = ["tests"]

[tool.ruff.lint.isort.sections]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.pytest.ini_options]
testpaths = ["tests"]